fix issue with custom handles or labels in legend
Changes to the legend function are made so that an issue with
custom labels or handles is removed. Beforehand, the kwargs would
interfere with the positional arguments and a warning was given.
Now, custom handles and / or labels kwargs override the default values
retrieved from axs[0] and are popped from the kwargs dict, before it
gets passed to the wrapped legend function.